refactor(coding-agent): share the zombie-process fixture and tighten comments
One spawnZombieProcess fixture replaces three copies of the perl fork
scaffold, and multi-line comment blocks collapse to one-line invariant
guards. No behavior or coverage change.
