+++ packages/coding-agent/test/child-process.test.ts
@@ -9,6 +9,7 @@ import {
 	signalProcessGroupOrProcess,
 	waitForChildProcess,
 } from "../src/utils/child-process.js";
+import { spawnZombieProcess } from "./fixtures/zombie-process.js";
 
 describe("waitForChildProcess", () => {
 	it("reports signaled already-exited children as failures", async () => {
@@ -37,65 +38,25 @@ describe("process liveness", () => {
 	});
 
 	it.skipIf(process.platform === "win32")("treats a zombie process as dead", async () => {
-		const parent = spawn(
-			"perl",
-			["-e", '$| = 1; my $pid = fork(); if ($pid) { print "$pid\\n"; sleep 30 } else { exit 0 }'],
-			{ stdio: ["ignore", "pipe", "ignore"] },
-		);
+		const { zombiePid, dispose } = await spawnZombieProcess();
 		try {
-			const zombiePid = await new Promise<number>((resolvePid, rejectPid) => {
-				let output = "";
-				const timer = setTimeout(() => rejectPid(new Error("Timed out waiting for the zombie pid")), 5000);
-				parent.stdout.on("data", (chunk: Buffer) => {
-					output += chunk.toString();
-					const parsed = Number.parseInt(output.trim(), 10);
-					if (Number.isInteger(parsed) && parsed > 0) {
-						clearTimeout(timer);
-						resolvePid(parsed);
-					}
-				});
-			});
-			const deadline = Date.now() + 5000;
-			while (!isZombieProcess(zombiePid) && Date.now() < deadline) {
-				await new Promise((resolveDelay) => setTimeout(resolveDelay, 25));
-			}
 			expect(isZombieProcess(zombiePid)).toBe(true);
 			expect(isProcessAlive(zombiePid)).toBe(false);
 		} finally {
-			parent.kill("SIGKILL");
+			dispose();
 		}
 	});
 
 	it.skipIf(process.platform === "win32")("does not let an unreaped zombie block group-stop completion", async () => {
-		const parent = spawn(
-			"perl",
-			["-e", '$| = 1; my $pid = fork(); if ($pid) { print "$pid\\n"; sleep 30 } else { setpgrp(0, 0); exit 0 }'],
-			{ stdio: ["ignore", "pipe", "ignore"] },
-		);
+		// setpgrp makes the zombie its group's only member: the group exists, but
+		// a stop waiting on it must complete because nothing is left running.
+		const { zombiePid, dispose } = await spawnZombieProcess("setpgrp(0, 0);");
 		try {
-			const zombiePid = await new Promise<number>((resolvePid, rejectPid) => {
-				let output = "";
-				const timer = setTimeout(() => rejectPid(new Error("Timed out waiting for the zombie pid")), 5000);
-				parent.stdout?.on("data", (chunk: Buffer) => {
-					output += chunk.toString();
-					const parsed = Number.parseInt(output.trim(), 10);
-					if (Number.isInteger(parsed) && parsed > 0) {
-						clearTimeout(timer);
-						resolvePid(parsed);
-					}
-				});
-			});
-			const deadline = Date.now() + 5000;
-			while (!isZombieProcess(zombiePid) && Date.now() < deadline) {
-				await new Promise((resolveDelay) => setTimeout(resolveDelay, 25));
-			}
 			expect(isZombieProcess(zombiePid)).toBe(true);
-			// The zombie is its group's only member: the group exists, but a stop
-			// waiting on it must complete because nothing is left running.
 			expect(processGroupExists(zombiePid)).toBe(true);
 			expect(processGroupHasLiveMember(zombiePid)).toBe(false);
 		} finally {
-			parent.kill("SIGKILL");
+			dispose();
 		}
 	});
